refactor(#5): extract account detail strategy and buying-center hooks
Move strategy save/fields and stakeholder/contact CRUD plus role updates into dedicated hooks so AccountDetailClient stays a thin tab composer.
